Euler angle visualisation - WIP
Allow user to visualise Euler 'disks' and choose between different Tensor ordering conventions.

TODO: The Euler angles on the canvas don't always match the equivalent set computed directly from the tensors. This needs to be fixed in the corresponding version of crystvis-js

TODO: the UI is functioning more or less correctly but needs streamlining.